Change LanesInUse to int64_t (#244)
LanesInUse is changed from size_t to int64_t to cather for
unknown case, this PR would align pldm changes to the new PDI
change.

+++ host-bmc/dbus/pcie_device.cpp
@@ -17,13 +17,13 @@ auto PCIeDevice::generationInUse(Generations value) -> Generations
         PCIeDevice::generationInUse(value);
 }
 
-size_t PCIeDevice::lanesInUse() const
+int64_t PCIeDevice::lanesInUse() const
 {
     return sdbusplus::xyz::openbmc_project::Inventory::Item::server::
         PCIeDevice::lanesInUse();
 }
 
-size_t PCIeDevice::lanesInUse(size_t value)
+int64_t PCIeDevice::lanesInUse(int64_t value)
 {
     return sdbusplus::xyz::openbmc_project::Inventory::Item::server::
         PCIeDevice::lanesInUse(value);
